requirements.txt
# SPDX-FileCopyrightText: 2022 Alec Delaney, for Adafruit Industries
#
# SPDX-License-Identifier: MIT

boto3~=1.34
click~=8.0
click-spinner~=0.1
packaging~=23.2
psutil~=5.9
pyyaml~=6.0
requests~=2.31
boto3-stubs[essential]~=1.34
typing-extensions~=4.0
